Prints para detectar problema

+++ tp/common/GrassTrackPiece.h
@@ -16,12 +16,12 @@ private:
   // float y;
 
 public:
-  GrassTrackPiece(float x, float y, int track_type, std::vector<size_t> size);
+  GrassTrackPiece(float x, float y, int track_type, std::vector<float> size);
   // void Place(b2World& world, b2Vec2 position);
   virtual void updateCarCounter(Car& car) override;
   const bool isCarOverMe(Car& car);
   const std::vector<float> GetPosition();
-  const std::vector<size_t> GetSize();
+  const std::vector<float> GetSize();
   const int getTrackType();
 
 };
